generator/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
python-docx==1.1.2
docxtpl==0.16.0
alembic==1.13.1
